feat: added new functional

+++ controllers/card.js
@@ -1,8 +1,9 @@
 const Card = require('../models/card');
 const { ValidationError } = require('../errors/ValidationError');
 const { NotFoundError } = require('../errors/NotFoundError');
+const { DeclinePermission } = require('../errors/DeclinePermission');
 
-const getCard = (req, res, next) => {
+const getCard = (_, res, next) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
     .catch((err) => {
@@ -13,27 +14,42 @@ const getCard = (req, res, next) => {
 const createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.status(201).send({ data: card }))
+    .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
         next(new ValidationError('400 - Переданы некорректные данные в методы создания карточки'));
-      } else {
-        next(err);
+        return;
       }
+      next(err);
     });
 };
 
 const deleteCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
+      const owner = card.owner.toHexString();
       if (!card) {
-        next(new NotFoundError('404 - Карточка с указанным _id не найдена.'));
+        next(new NotFoundError('404 - Карточка с указанным id не найдена'));
+      } else if (owner === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .orFail(new Error('NoValidId'))
+          .then((cardDeleted) => res.send(cardDeleted))
+          .catch((err) => {
+            if (err.message === 'NoValidId') {
+              next(new NotFoundError('404 - Карточка с указанным _id не найдена.'));
+            } else {
+              next(err);
+            }
+          });
+      } else {
+        next(new DeclinePermission('403 — Попытка удалить чужую карточку'));
       }
-      res.send({ data: card });
     })
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        next(new ValidationError('400 - Отправлены некорректные данные'));
+      if (err.name === 'CastError') {
+        next(new ValidationError('400 —  Карточка с указанным _id не найдена'));
+      } else if (err.name === 'TypeError') {
+        next(new NotFoundError('404 - Удаление карточки с несуществующим в БД id'));
       } else {
         next(err);
       }
@@ -44,16 +60,17 @@ const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        res.status(404).send({ message: 'Передан несуществующий _id карточки.' });
+        next(new NotFoundError('404 - Передан несуществующий _id карточки.'));
+        return;
       }
       res.send({ data: card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
         next(new ValidationError('400 - Переданы некорректные данные для постановки/снятии лайка.'));
-      } else {
-        next(err);
+        return;
       }
+      next(err);
     });
 };
 
@@ -61,16 +78,17 @@ const dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        res.status(404).send({ message: 'Передан несуществующий _id карточки.' });
+        next(new NotFoundError('404 - Передан несуществующий _id карточки.'));
+        return;
       }
       res.send({ data: card });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
         next(new ValidationError('400 - Переданы некорректные данные для постановки/снятии лайка.'));
-      } else {
-        next(err);
+        return;
       }
+      next(err);
     });
 };
